[#17] Refactor: HeroSection 컴포넌트 분리
